feat: 新增 確認登入狀態後跳轉頁面功能
