+++ src/scripts/methods.js
@@ -6,3 +6,20 @@ export const pushMessage = (msg) => {
 };
 
 export const bsModal = (id) => new Modal(document.getElementById(id));
+
+export const setToken = (token) => {
+  const expries = new Date(Date.now() + 1000 * 60 * 30).toGMTString();
+  document.cookie = `nodeFinal=${token}; expires=${expries}`;
+};
+
+export const getToken = () => {
+  const token = document.cookie.replace(
+    /(?:(?:^|.*;\s*)nodeFinal\s*=\s*([^;]*).*$)|^.*$/,
+    "$1"
+  );
+  return token;
+};
+
+export const clearToken = () => {
+  document.cookie = "nodeFinal=;expires=;path=/";
+};
